mydjango14/requirements.txt
django~=3.2.0
django-debug-toolbar
pillow
django-imagekit